nvmf: namespace masking with attach/detach
This patch is an evolution of RFC:

https://review.spdk.io/gerrit/c/spdk/spdk/+/7821

It introduces namespace masking for NVMe-over-fabrics
targets by allowing to (dynamically) attach and detach
controllers to/from namespaces, cf. NVMe spec 1.4 - section 6.1.4.
Since SPDK only supports the dynamic controller model a new
controller is allocated on every fabric connect command.
This allows to attach/detach controllers of a specific
host NQN to/from a namespace. A host can only perform
operations to an active namespace. Inactive namespaces can
be listed (not supported by SPDK) but no additional
information can be retrieved:
"Unless otherwise noted, specifying an inactive NSID in a
command that uses the Namespace Identifier (NSID) field shall
cause the controller to abort the command with status
Invalid Field in Command" - NVMe spec 1.4 - section 6.1.5
Note that this patch does not implement the NVMe namespace
attachment command but allows to attach/detach via RPCs only.
To preserve current behavior all controllers are auto attached.
To not not auto attach controllers the nvmf_subsystem_add_ns
shall be called with "--no-auto-attach". We introduce two new
RPC calls:
- nvmf_ns_attach_ctrlr <subsysNQN> <NSID> <hostNQN> [-hot] [-cold]
- nvmf_ns_detach_ctrlr <subsysNQN> <NSID> <hostNQN> [-hot] [-cold]
If -hot is specified only currently active controllers which
hostNQN match the hostNQN specified are attached or detached
to/from the namespace. New connections/controllers are not
attached to the namespace. If -cold is specified only new
connections/controllers are attached. Currently connected
controllers of hosts are not attached/detached. If both
-hot or -cold are specified or none controllers are both
hot and cold attached/detached to/from a namespace.
The list in spdk_nvmf_ns is used to keep track of hostNQNs
which controllers should be attached on connect.
The active_ns array in spdk_nvmf_ctrlr is used for fast lookup
to check whether a NSID is active/inactive on command execution.

Changes to RFC:
- Address concerns of overhead of checking on IO path.
  If a namepsace is active is only checked where needed and
  validity checks for NSID is only done once.
- Allow to hot and cold attach/detach
- It is no longer possible to call the new RPCs without
  hostNQN. The auto attach bevahior cannot be changed
  after add_ns. We might add an additional RPC for this
  later.
- identify ns now correctly response with zeros for
  inactive namespace
- Address concurrency concerns: attaching/detaching
  now pauses the subsystem for changes
- ns change log is updated on attach/detach
- tests

Change-Id: I1c3a3c2bf3615fa76e7bfe3aaf08e2b82d6b4ec0
Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>