test/bdevio: fix bytes vs blocks => early return of test
blockdev_compare_and_write and blockdev_write_read check
if data_length is multiple of block size and silently
fail if it is not. There are multiple test that test
e.g. invalid size but the actual bdev API is
never called and tested because of the early return.
Remove test for valid block size and let the return
value of the bdev functions indicate failure like
intended.
Furthermore, no cmp&write test is ever executed because
data_length is set to 1 trying to indicate 1 block but
it is tested against valid block bytes and silently fails.
Also write and read helpers in cmp&write test expect bytes
not blocks. For consistency use bytes everywhere and only
convert to blocks when calling the cmp&write blocks API.
512bytes read/write test always silently failed for e.g.
aio bdev because block size was set to 2048. Change
block size to 512bytes.

Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>
Change-Id: I662094c2bcd08c0e9dc5573177a2e7a0edd275ce