bdev: if md is separate => md is required
Currently we allow to issue cmds without a md buffer (md == NULL)
when the bdev has md and md is separate. This is problematic for e.g.
the nvme bdev since there is no way to indicate that you want to
do io without md in a NVMe cmd. All io requires md if the ns is
formatted with md. And If the ns is formatted with md and the md
is separate one has to provide a MPTR.
This patch adds checks for the case when there is md and md is
separate to require a md_buf to be specified (!=NULL).
We distinguish the two cases:
1) no md or md is interleaved
2) md is separate

For 1) where 0 == NULL, 1 != NULL:
buf | md_buf | rc
------------------------
 0  |   0    | 0
 0  |   1    | -EINVAL
 1  |   0    | 0
 1  |   1    | -EINVAL

For 2) where 0 == NULL, 1 != NULL:
buf | md_buf | rc
------------------------
 0  |   0    | 0
 0  |   1    | -EINVAL
 1  |   0    | -EINVAL
 1  |   1    | 0

Signed-off-by: Jonas Pfefferle <pepperjo@japf.ch>
Change-Id: I5764c875bca6826bbbff8e0f5c0cc0ed1bafbc64